requirements.txt
beautifulsoup4~=4.13.3
python-dotenv~=1.0.1
requests~=2.32.3
playwright~=1.50.0
pytelegrambotapi~=4.26.0

fastapi~=0.115.6